Refactoring, fix, RAM savings
Refactoring: moved programming out of loop() because it'll probably be changed
to run at the push of a button, rather than when something's detected on the
bus.

Fix: averaging worked by dividing the sum of valid temperatures by the number
of sensors, but the number of valid temperatures can be lower than the number
of sensors. The number of valid temperatures is now counted independently.

RAM saving: lower the number of sensors supported from 64 to 40, to conserve
RAM; it was getting tight.